# docs/openapi_current.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "bin2nlp API",
        "description": "\n        ## Binary Decompilation & Multi-LLM Translation Service\n\n        Transform binary executables into human-readable analysis through advanced decompilation \n        and multi-provider LLM translation.\n\n        ### Key Features\n        - **Multi-format support**: PE, ELF, Mach-O binary analysis\n        - **Multi-LLM providers**: OpenAI, Anthropic, Google Gemini integration\n        - **Scalable architecture**: PostgreSQL database, file storage, async processing\n        - **Production ready**: Open access, monitoring, observability\n\n        ### Open Access\n        All endpoints are freely accessible without authentication requirements.\n        This service is designed for open binary analysis and research.\n\n        ### Support\n        - Documentation: [GitHub Repository](https://github.com/example/bin2nlp)\n        - Issues: [Issue Tracker](https://github.com/example/bin2nlp/issues)\n        ",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "bin2nlp Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.bin2nlp.com",
...
        }
    ],
    "paths": {
...
        "/api/v1/decompile": {
            "post": {
                "tags": [
                    "decompilation"
                ],
                "summary": "Submit Decompilation Job",
                "description": "Submit a binary file for decompilation and LLM translation.\n\nUploads a binary file and starts async decompilation with optional\nLLM translation. Returns a job ID for tracking progress.\n\nArgs:\n    file: Binary file to decompile (PE, ELF, Mach-O, etc.)\n    analysis_depth: Decompilation depth (basic, standard, comprehensive)\n    llm_provider: LLM provider for translation (openai, anthropic, gemini)\n    translation_detail: Translation detail level (basic, standard, detailed)\n\nReturns:\n    Job information with tracking ID",
                "operationId": "submit_decompilation_job_api_v1_decompile_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_submit_decompilation_job_api_v1_decompile_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/decompile/{job_id}": {
            "get": {
                "tags": [
                    "decompilation"
                ],
                "summary": "Get Decompilation Result",
                "description": "Get decompilation job status and results.\n\nReturns job status, progress information, and complete results\nwhen processing is finished.",
                "operationId": "get_decompilation_result_api_v1_decompile__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Id"
                        }
                    },
                    {
                        "name": "include_raw_data",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Include Raw Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "decompilation"
                ],
                "summary": "Cancel Decompilation Job",
                "description": "Cancel a pending or in-progress decompilation job.\n\nOnly jobs that are pending or in early processing stages can be cancelled.",
                "operationId": "cancel_decompilation_job_api_v1_decompile__job_id__delete",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Body_submit_decompilation_job_api_v1_decompile_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "analysis_depth": {
                        "type": "string",
                        "title": "Analysis Depth",
                        "default": "standard"
                    },
                    "llm_provider": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Provider"
                    },
                    "llm_model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Model"
                    },
                    "llm_endpoint_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Endpoint Url"
                    },
                    "llm_api_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Api Key"
                    },
                    "translation_detail": {
                        "type": "string",
                        "title": "Translation Detail",
                        "default": "standard"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_submit_decompilation_job_api_v1_decompile_post"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}